sa3: review fixes, helper relocation, and family-aware boot preflight
Review-driven correctness/cleanup on the SA3 backend family:
- adapter: warp the shift_alpha schedule on the [0,1]-normalized values
  (was warping sigma_max-scaled values + re-pinning only index 0), which
  could invert the first step at partial denoise. Regression test added.
- backend: guard the conditioning control state (cond/bundle/history +
  schedule builder/cache) with a lock so a command-thread prompt swap
  can't race the runner thread's reads (cond_history append+truncate vs
  iterate; non-atomic bundle/builder swaps).
- trt: key the DiT staging cache on a held bundle reference, not id(),
  which could stale-hit after object reuse and run stale conditioning.
- build: single-source the DiT profile matrix (dry-run vs --all).
- stream helpers: collapse a copy-paste-dead chunk_midpoint_shift branch.

Structure:
- move the load-bearing SA3 runtime helpers (conditioning, source encode,
  SAME windowed decode) into acestep/engine/sa3_stream_helpers.py so
  production no longer imports them out of scripts/ over a front-injected
  sys.path; the spike module re-exports them for its test/demo ringbuffer.

Serving:
- static_site: a malformed AUTO-DISCOVERED repo demo manifest is now
  warned and skipped instead of SystemExit-ing the shared host; explicit
  --demo paths still fail loudly.
- server: family-dispatch the boot preflight. The ACE preflight checks
  the acestep checkpoint/engine tree, which false-fails for the sa3
  family (whose weights/engines live under <models>/sa3/). sa3 gets a
  light readiness check (model safetensors + vendored source; engines
  are non-fatal, eager fallback). sa3_checkpoint_dir single-sourced.
